+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Remove cloudfront secret dependency from reconcilers.
+
 ## [0.9.0] - 2023-06-08
 
 ### Changed